v0.2.18: message UUID and component sequence on message metadata
In this commit, UUIDs are added to the component metadata when a message is first
created. This allows for better tracking of the message traversing the system.
Also, the sequence of components handling a message is recorded.

The effects of this change can be seen in the example application, where some
messages routes are printed to the browser console or the terminal for the backend
application. I was undecided about the best way to record the sequence, so I created
two different formats. Try it out and let me know what you think. You'll need to
install the systems-toolbox locally for now as this WIP isn't published yet.